Make errorbar3d property cycling match 2D.
As with 2D before #17930, it would not cycle if a color were specified.
However, this does not match `plot`, which does not advance the cycle
only if _all_ properties in the cycle are specified. Notably, this means
if your property cycle was for line style, specifying a color would
ignore the cycle in `errorbar`, but not in `plot`.

This is a 3D version of 149e7fbf281541039b1b5a79d6efa7ab99f582f0 and
0782c743c0823bc823c16c3d66cea0d42df5fe93.