polar: Update Axes patch as part of transform update
Instead of generating a throwaway `Wedge` class to calculate the
boundaing box for sizing, re-use the one that is the `Axes` patch
(unless it was never set). This also removes the redundant
re-calculation to resize the `Axes` patch during draw.

This does in turn require tweaking the transform, as the calculation
uses a slightly different domain, but it should be more robust to
changes in the bbox calculation.