Move cibuildwheel configuration to pyproject.toml
This allows developers to test out the build locally. With this, I was
able to fix the test command by setting `PIP_PREFER_BINARY` to avoid
re-building Pillow, which accidentally dropped manylinux2014 wheels in
the latest release: https://github.com/python-pillow/Pillow/issues/9057

Note also that we previously set `CIBW_AFTER_BUILD`, but this doesn't
seem to be a valid setting. Thus I have dropped the `twine check`, which
only tests the `README` rendering and so checking the sdist is
sufficient. Additionally, I have commented out the license check, as we
cannot do multiple licenses with meson-python without PEP639 (#28982).